♻️ Move RSS sniffer to its own file
